// frontend/src/types/settings.ts
export interface ISettingsCompany {
  kode: string;
  nama: string;
  alamat: string;
  kota: string;
  kodepos: string;
  telepon: string;
  faks: string;
  email: string;
  logo: string;
  nama_pimpinan: string;
  jabatan_pimpinan: string;
  ttd_pimpinan: string;
}

export interface ISettingsNumbering {
  alias?: string;
  pemisah?: number;
  format1?: number;
  format2?: number;
  format3?: number;
  format4?: number;
  contoh?: string;
  reset?: number;
  noseri?: string;
  inicab?: string;
  digitNomor?: string;
  bkk?: number;
  nobkk?: string;
  bkm?: number;
  nobkm?: string;
  bbm?: number;
  nobbm?: string;
  bbk?: number;
  nobbk?: string;
  bmm?: number;
  nobmm?: string;
  bjk?: number;
  nobjk?: string;
  pjl?: number;
  nopjl?: string;
  pbl?: number;
  nopbl?: string;
  bppb?: number;
  nobppb?: string;
  bpsb?: number;
  nobpsb?: string;
  bbp?: number;
  nobbp?: string;
  bpb?: number;
  nobpb?: string;
  sprk?: number;
  nosprk?: string;
  bsprk?: number;
  nobsprk?: string;
  ppl?: number;
  noppl?: string;
  bpl?: number;
  nobpl?: string;
  po?: number;
  nopo?: string;
  bpo?: number;
  nobpo?: string;
  bp?: number;
  nobp?: string;
  bpsprk?: number;
  noins?: string;
  ins?: number;
  nokns?: string;
  kns?: number;
  norpb?: string;
  rpb?: number;
  nospg?: string;
  spg?: number;
  noopn?: string;
  opn?: number;
  nokms?: string;
  kms?: number;
  norbpb?: string;
  rbpb?: number;
  noenq?: string;
  enq?: number;
  nocr?: string;
  cr?: number;
  noso?: string;
  so?: number;
  norkl?: string;
  rkl?: number;
  nosc?: string;
  sc?: number;
  nospp?: string;
  spp?: number;
  nospb?: string;
  spb?: number;
  norspb?: string;
  rspb?: number;
  noinv?: string;
  inv?: number;
  nopnj?: string;
  pnj?: number;
  notrc?: string;
  trc?: number;
  noship?: string;
  ship?: number;
  notbj?: string;
  tbj?: number;
  norbj?: string;
  rbj?: number;
  notrs?: string;
  trs?: number;
  akm?: number;
  norpj?: string;
  rpj?: number;
  nobhn?: string;
  bhn?: number;
  spk?: string;
  rbp?: string;
  pr?: string;
  dn?: string;
  opbj?: string;
  kmbj?: string;
  invc?: string;
  noInvc?: string;
  spr?: string;
  noSpr?: string;
  kn?: string;
  noKn?: string;
  hpd?: string;
  pbs?: string;
  rpbs?: string;
  pos?: string;
  ht?: string;
  pt?: string;
  tt?: string;
  krs?: string;
  avt?: string;
}